GHI-543 extracted common code into separate file
